pyproject.toml
[tool.poetry]
name = "secure-url"
version = "0.1.0"
description = ""
authors = ["Bartek Krzys <bartek.krzys@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
psycopg2 = "^2.8.6"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
